requirements.txt
numpy
pip-review
pytest
pyyaml
sklearn
tensorflow
